+++ src/main/java/itu/GreenField/dto/EvolutionVenteDto.java
@@ -1,8 +1,9 @@
 package itu.greenfield.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record EvolutionVenteDto(
     LocalDate date,
-    Double totalVentes
+    BigDecimal totalVentes
 ) {}
